% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_sample.R
\name{accept_reject_psu}
\alias{accept_reject_psu}
\title{Accept or reject sampling units against population density of potential
sampling units}
\usage{
accept_reject_psu(
  x,
  svy,
  psu = "psu",
  match = NULL,
  pop = "pop",
  verbose = TRUE,
  show_plot = TRUE,
  save_plot = FALSE
)
}
\arguments{
\item{x}{A data.frame of all potential primary sampling units (PSUs) to
sample from each uniquely identified and with corresponding populations.}

\item{svy}{A data.frame of survey data drawn via probability proportional
to population size (PPS) approach from all the potential PSUs in \emph{x}.}

\item{psu}{A single character value or vector of values corresponding to
the variable name/s for the primary sampling unit in \emph{x} and \emph{svy}.
Default is \emph{"psu"}.}

\item{match}{A single character value corresponding to the variable name
for the primary sampling unit in \emph{x} that matches the primary sampling
unit in \emph{svy}. Default is NULL which indicates that primary sampling unit
in \emph{x} is the same as primary sampling unit in \emph{svy}.}

\item{pop}{A single character value corresponding to the variable name for
the population figures in \emph{x}. Default is \emph{"pop"}.}

\item{verbose}{Logical. Should text of sample acceptance or rejection be
printed/shown? Default to TRUE.}

\item{show_plot}{Logical. Should plot of sample acceptance or rejection be
shown? Default to TRUE.}

\item{save_plot}{Logical. Should plot of sample acceptance or rejection be
saved? Default to FALSE. If set to TRUE when \emph{show_plot} is FALSE, no plot
will be saved.}
}
\value{
A data.frame drawn from \emph{svy} that contains the primary sampling
units selected by the acceptance and rejection sampling.
}
\description{
Accept or reject sampling units against population density of potential
sampling units
}
\examples{
accept_reject_psu(
  x = village_list,
  svy = sample_data,
  psu = c("id", "psu"),
  match = "cluster",
  pop = "population",
  verbose = FALSE,
  show_plot = FALSE
 )

}
