% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sample_systematic.R
\name{calculate_sampling_interval}
\alias{calculate_sampling_interval}
\title{Calculate sampling interval}
\usage{
calculate_sampling_interval(
  n_sample,
  n_total,
  rounding = c("floor", "ceiling")
)
}
\arguments{
\item{n_sample}{Number of sample required}

\item{n_total}{Total number to sample from}

\item{rounding}{Should the result be rounded up (ceiling) or rounded down
(floor)? Default is rounded down.}
}
\value{
An integer value representing the sampling interval.
}
\description{
Calculate sampling interval
}
\examples{
calculate_sampling_interval(30, 200)

}
\author{
Ernest Guevarra
}
