% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_sample_psm.R
\name{create_sample_psm}
\alias{create_sample_psm}
\title{Select sampling units from a set of potential sampling units using propensity
score matching (PSM)}
\usage{
create_sample_psm(
  x,
  svy,
  psu = "psu",
  match = NULL,
  pop = "pop",
  sampling_type = c("simple", "systematic")
)
}
\arguments{
\item{x}{A data.frame of all potential primary sampling units (PSUs) to
sample from each uniquely identified and with corresponding populations.}

\item{svy}{A data.frame of survey data drawn via probability proportional
to population size (PPS) approach from all the potential PSUs in \emph{x}.}

\item{psu}{A single character value or vector of values corresponding to
the variable name/s for the primary sampling unit in \emph{x} and \emph{svy}.
Default is \emph{"psu"}.}

\item{match}{A single character value corresponding to the variable name
for the primary sampling unit in \emph{x} that matches the primary sampling
unit in \emph{svy}. Default is NULL which indicates that primary sampling unit
in \emph{x} is the same as primary sampling unit in \emph{svy}.}

\item{pop}{A single character value corresponding to the variable name for
the population figures in \emph{x}. Default is \emph{"pop"}.}

\item{sampling_type}{Which sampling type to use. Choice between \emph{simple
random sample (simple)} or \emph{systematic sample (systematic)}. Default is
\emph{simple}.}
}
\value{
A data.frame drawn from \emph{svy} that contains the primary sampling
units selected by propensity score matching.
}
\description{
Select sampling units from a set of potential sampling units using propensity
score matching (PSM)
}
\examples{
create_sample_psm(
  x = village_list,
  svy = sample_data,
  psu = c("id", "psu"),
  match = "cluster",
  pop = "population"
)

}
\author{
Mark Myatt and Ernest Guevarra
}
