% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sample_systematic.R
\name{get_sample_systematic}
\alias{get_sample_systematic}
\title{Draw a systematic sample from a dataset of all potential samples}
\usage{
get_sample_systematic(x, n_sample, rounding = c("floor", "ceiling"))
}
\arguments{
\item{x}{A data.frame of all potential samples}

\item{n_sample}{Number of sample required}

\item{rounding}{Should the result be rounded up (ceiling) or rounded down
(floor)? Default is rounded down.}
}
\value{
A data.frame drawn from \emph{x} that contains the sample selected
systematically
}
\description{
Draw a systematic sample from a dataset of all potential samples
}
\examples{
get_sample_systematic(x = village_list, n_sample = 30)

}
\author{
Ernest Guevarra
}
