% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_sample_psm.R
\name{get_sampling_clusters}
\alias{get_sampling_clusters}
\title{Select sampling units from a set of potential sampling units using propensity
score matching (PSM)}
\usage{
get_sampling_clusters(
  x,
  psu = NULL,
  pop = NULL,
  sampling_type = c("simple", "systematic")
)
}
\arguments{
\item{x}{A data.frame of all potential primary sampling units (PSUs) to
sample from each uniquely identified and with corresponding populations.}

\item{psu}{A single character value or vector of values corresponding to
the variable name/s for the primary sampling unit in \emph{x} and \emph{svy}.
Default is \emph{"psu"}.}

\item{pop}{A single character value corresponding to the variable name for
the population figures in \emph{x}. Default is \emph{"pop"}.}

\item{sampling_type}{Which sampling type to use. Choice between \emph{simple
random sample (simple)} or \emph{systematic sample (systematic)}. Default is
\emph{simple}.}
}
\value{
A data.frame drawn from \emph{x} that contains the primary sampling
units selected by propensity score matching.
}
\description{
Select sampling units from a set of potential sampling units using propensity
score matching (PSM)
}
\examples{
get_sampling_clusters(
  x = village_list,
  psu = "cluster",
  pop = "population"
)

}
\author{
Mark Myatt and Ernest Guevarra
}
