% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_data}
\alias{sample_data}
\title{Sample dataset from villages selected from full list via probability
proportional to population size}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 407 rows and 11 columns.
}
\usage{
sample_data
}
\description{
A tibble with 407 rows and 11 columns:
}
\details{
\tabular{ll}{
   \strong{Variable} \tab \strong{Description} \cr
   \emph{surveydate} \tab Date of survey \cr
   \emph{psu} \tab Primary sampling unit identifier \cr
   \emph{sex} \tab Sex of child \cr
   \emph{birthdate} \tab Date of birth of child \cr
   \emph{age} \tab Age of child in months \cr
   \emph{weight} \tab Weight of child in kilograms \cr
   \emph{height} \tab height of child in centimetres \cr
   \emph{oedema} \tab Does the child have oedema? 0 = FALSE; 1 = TRUE \cr
   \emph{muac} \tab Mid-upper arm circumference in millimetres \cr
   \emph{measure} \tab Was the child's height or length measured? l = length; h = height \cr
   *clothes( \tab Was the child wearing clothes when measured? y = TRUE; n = FALSE) \cr
}
}
\examples{
sample_data

}
\keyword{datasets}
