% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{village_list}
\alias{village_list}
\title{Full village list with population size and cluster identifier for villages
selected by probability proportional to population size}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 73 rows and 4 columns.
}
\source{
Data from a SMART survey conducted in Cameroon in 2020
}
\usage{
village_list
}
\description{
A tibble with 73 rows and 4 columns:
}
\details{
\tabular{ll}{
   \strong{Variable} \tab \strong{Description} \cr
   \emph{id} \tab Village identifier \cr
   \emph{village} \tab Village name \cr
   \emph{population} \tab Population size \cr
   \emph{cluster} \tab Cluster identifier for selected villages \cr
}
}
\examples{
village_list

}
\keyword{datasets}
