% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveysampler.R
\docType{package}
\name{zscorer}
\alias{surveysampler}
\alias{surveysampler-package}
\title{Survey Sampling and Analysis Tools}
\description{
Studies that evaluate survey sampling and analysis approaches require varied
techniques and methods not found in a single package in R. This package
provides utilities that aid and simplify these approaches to enable
streamlined assessment and comparison of different survey sampling and
analysis techniques.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ernestguevarra/surveysampler}
  \item Report bugs at \url{https://github.com/ernestguevarra/surveysampler/issues}
}

}
\author{
\strong{Maintainer}: Ernest Guevarra \email{ernest@guevarra.io} (\href{https://orcid.org/0000-0002-4887-4415}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Mark Myatt \email{mark@brixtonhealth.com} (\href{https://orcid.org/0000-0003-1119-1474}{ORCID}) [copyright holder]
}

}
\keyword{internal}
